Fix KCoreAddons build on Clang+ARM
We have a set of reverse-engineered literals and one of them we're
specifying it as unsigned (the rest are smaller, so it doesn't make a
difference), even though the API returns a signed int.

Tries to address the issue pointed out by Rohan in the mailing list.

REVIEW: 124150

+++ src/lib/io/kfilesystemtype.cpp
@@ -99,7 +99,7 @@ KFileSystemType::Type determineFileSystemTypeImpl(const QByteArray &path)
         //qDebug() << path << errno << strerror(errno);
         return KFileSystemType::Unknown;
     }
-    switch (buf.f_type) {
+    switch (static_cast<unsigned long>(buf.f_type)) {
     case NFS_SUPER_MAGIC:
     case AUTOFS_SUPER_MAGIC:
     case AUTOFSNG_SUPER_MAGIC:
